PLUGIN-11 Address comments
