+++ src/main/java/io/cdap/plugin/couchbase/source/N1qlQueryRowInputFormatProvider.java
@@ -20,7 +20,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import io.cdap.cdap.api.data.batch.InputFormatProvider;
-import io.cdap.plugin.couchbase.CouchbaseConfig;
+import io.cdap.plugin.couchbase.CouchbaseSourceConfig;
 
 import java.util.Map;
 
@@ -33,7 +33,7 @@ public class N1qlQueryRowInputFormatProvider implements InputFormatProvider {
 
   private final Map<String, String> conf;
 
-  public N1qlQueryRowInputFormatProvider(CouchbaseConfig config) {
+  public N1qlQueryRowInputFormatProvider(CouchbaseSourceConfig config) {
     this.conf = new ImmutableMap.Builder<String, String>()
       .put(PROPERTY_CONFIG_JSON, gson.toJson(config))
       .build();
